Backed out 5 changesets (bug 1890092, bug 1888683) for causing build bustages & crashes CLOSED TREE
Backed out changeset d43df75c14de (bug 1888683)
Backed out changeset b4267327c7dd (bug 1888683)
Backed out changeset 11892428e2dc (bug 1890092)
Backed out changeset 42598e432246 (bug 1888683)
Backed out changeset 372016bf3614 (bug 1888683)

UltraBlame original commit: 7da2e2d38581a7abe632df2bc54c098d53587057

+++ toolkit/library/rust/shared/lib.rs
@@ -174,7 +174,7 @@ pub unsafe extern "C" fn debug_log(target: *const c_char, message: *const c_char
 
 #[no_mangle]
 pub extern "C" fn uniffi_rustbuffer_alloc(
-    size: u64,
+    size: i32,
     call_status: &mut uniffi::RustCallStatus,
 ) -> uniffi::RustBuffer {
     uniffi::uniffi_rustbuffer_alloc(size, call_status)
